Task queueing and dequeuing work, still fixing threading cleanup
