+++ hyrex/__init__.py
@@ -1,2 +1,4 @@
-from .app import DispatcherType, Hyrex
-from .task_registry import TaskRegistry
+# from .app import DispatcherType, Hyrex
+# from .task_registry import TaskRegistry
+from .hyrex_registry import HyrexRegistry
+from .worker import HyrexWorker
