+++ hyrex/dispatcher/__init__.py
@@ -1,4 +1,2 @@
-from .dispatcher import DequeuedTask, Dispatcher
-from .platform_dispatcher import PlatformDispatcher
-from .postgres_dispatcher import PostgresDispatcher
-from .postgres_lite_dispatcher import PostgresLiteDispatcher
+from .dispatcher_provider import get_dispatcher
+from .dispatcher import Dispatcher, DequeuedTask
