+++ hyrex/dispatcher/dispatcher.py
@@ -8,28 +8,28 @@ from hyrex import constants
 from hyrex.models import HyrexTask, StatusEnum
 
 
-class SingletonMeta(ABCMeta):
-    _instances = {}
-    _init_args = {}
-    _lock = threading.Lock()
+# class SingletonMeta(ABCMeta):
+#     _instances = {}
+#     _init_args = {}
+#     _lock = threading.Lock()
 
-    def __call__(cls, *args, **kwargs):
-        # If the instance does not exist, create it and store args
-        if cls not in cls._instances:
-            with cls._lock:
-                if cls not in cls._instances:
-                    instance = super().__call__(*args, **kwargs)
-                    cls._instances[cls] = instance
-                    cls._init_args[cls] = (args, kwargs)
-        else:
-            # Check if the new args differ from the initial args
-            if (args, kwargs) != cls._init_args[cls]:
-                raise ValueError(
-                    f"Singleton instance already created with "
-                    f"different arguments: {cls._init_args[cls]} vs {args, kwargs}"
-                )
+#     def __call__(cls, *args, **kwargs):
+#         # If the instance does not exist, create it and store args
+#         if cls not in cls._instances:
+#             with cls._lock:
+#                 if cls not in cls._instances:
+#                     instance = super().__call__(*args, **kwargs)
+#                     cls._instances[cls] = instance
+#                     cls._init_args[cls] = (args, kwargs)
+#         else:
+#             # Check if the new args differ from the initial args
+#             if (args, kwargs) != cls._init_args[cls]:
+#                 raise ValueError(
+#                     f"Singleton instance already created with "
+#                     f"different arguments: {cls._init_args[cls]} vs {args, kwargs}"
+#                 )
 
-        return cls._instances[cls]
+#         return cls._instances[cls]
 
 
 class DequeuedTask(BaseModel):
@@ -38,7 +38,8 @@ class DequeuedTask(BaseModel):
     args: dict
 
 
-class Dispatcher(ABC, metaclass=SingletonMeta):
+# class Dispatcher(ABC, metaclass=SingletonMeta):
+class Dispatcher(ABC):
     @abstractmethod
     def enqueue(
         self,
